Refactor: resolve critical bugs, security vulnerabilities, and code smells
- Security: Restrict CORS wildcard in api-gateway, secure Actuator endpoints in user-service, and use env variables for API URL in AuthContext.
- Bugfix: Add timeout to Kafka .get() blocking calls in OutboxScheduler and TradeOutboxScheduler to prevent thread starvation.
- Bugfix: Safely handle empty Redis streams in DynamicAuthorizationFilter using onErrorResume.
- Refactor (Domain): Migrate monetary fields (pnl, fees, amount, etc.) from primitive double to BigDecimal in Trade and BillingTransaction entities to prevent financial floating-point errors.
- Refactor (Architecture): Split monolithic AnalyticsServiceImpl into domain-specific services (AnalyticsOverviewService, AnalyticsPerformanceService, AnalyticsRiskService, AnalyticsMoodService) to strictly follow Single Responsibility Principle (SRP).
- Code Quality: Replace raw HashMap return types with GalleryItemDto.
- Saga: Implement SagaUserRoleConsumerService to properly update local AnalyticsConfig projection upon Premium upgrades.

+++ server-java/src/main/java/com/conarum/tradingjournal/domain/analytics/dto/AnalyticsResponseDto.java
@@ -6,6 +6,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class AnalyticsResponseDto {
@@ -17,22 +18,22 @@ public class AnalyticsResponseDto {
     @AllArgsConstructor
     public static class Overview {
         private int totalTrades;
-        private double totalPnl;
-        private double totalFees;
+        private BigDecimal totalPnl;
+        private BigDecimal totalFees;
         private int winners;
         private int losers;
         private int breakeven;
-        private double winRate;
-        private double avgWin;
-        private double avgLoss;
-        private double grossProfit;
-        private double grossLoss;
-        private double profitFactor;
-        private double expectancy;
-        private double sharpeRatio;
-        private double maxDrawdown;
-        private double currentDrawdown;
-        private double avgRR;
+        private BigDecimal winRate;
+        private BigDecimal avgWin;
+        private BigDecimal avgLoss;
+        private BigDecimal grossProfit;
+        private BigDecimal grossLoss;
+        private BigDecimal profitFactor;
+        private BigDecimal expectancy;
+        private BigDecimal sharpeRatio;
+        private BigDecimal maxDrawdown;
+        private BigDecimal currentDrawdown;
+        private BigDecimal avgRR;
         private int tradingDays;
     }
 
@@ -44,9 +45,9 @@ public class AnalyticsResponseDto {
     public static class ByCategory {
         private String category; // day, instrument, side, tag, playbookId
         private int trades;
-        private double pnl;
-        private double winRate;
-        private double avgPnl;
+        private BigDecimal pnl;
+        private BigDecimal winRate;
+        private BigDecimal avgPnl;
         
         // Specific fields
         private Integer dayIndex;
@@ -75,9 +76,9 @@ public class AnalyticsResponseDto {
     public static class Risk {
         private int tradesWithSL;
         private int tradesWithTP;
-        private double avgRR;
-        private double tpHitRate;
-        private double slHitRate;
+        private BigDecimal avgRR;
+        private BigDecimal tpHitRate;
+        private BigDecimal slHitRate;
         private List<Object> rrData;
         private List<Object> drawdownCurve;
     }
@@ -90,9 +91,9 @@ public class AnalyticsResponseDto {
     public static class Mood {
         private String mood;
         private int days;
-        private double totalPnl;
-        private double avgPnlPerDay;
+        private BigDecimal totalPnl;
+        private BigDecimal avgPnlPerDay;
         private int totalTrades;
-        private double winRate;
+        private BigDecimal winRate;
     }
 }
