+++ server-java/src/main/java/com/conarum/tradingjournal/domain/trade/dto/TradeFilterDto.java
@@ -3,6 +3,8 @@ package com.conarum.tradingjournal.domain.trade.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class TradeFilterDto {
@@ -13,8 +15,8 @@ public class TradeFilterDto {
     private String tag;
     private String dateFrom;
     private String dateTo;
-    private Double pnlMin;
-    private Double pnlMax;
+    private BigDecimal pnlMin;
+    private BigDecimal pnlMax;
     private String search;
     private String playbookId;
     private Integer rating;
